+++ src/gameobjects/container/ContainerCanvasRenderer.js
@@ -44,6 +44,14 @@ var ContainerCanvasRenderer = function (renderer, container, interpolationPercen
         transformMatrix.applyITRS(container.x, container.y, container.rotation, container.scaleX, container.scaleY);
     }
 
+    var containerHasBlendMode = (container.blendMode !== -1);
+
+    if (!containerHasBlendMode)
+    {
+        //  If Container is SKIP_TEST then set blend mode to be Normal
+        renderer.setBlendMode(0);
+    }
+
     var alpha = container._alpha;
     var scrollFactorX = container.scrollFactorX;
     var scrollFactorY = container.scrollFactorY;
@@ -58,19 +66,26 @@ var ContainerCanvasRenderer = function (renderer, container, interpolationPercen
         }
 
         var childAlpha = child._alpha;
+        var childBlendMode = child._blendMode;
         var childScrollFactorX = child.scrollFactorX;
         var childScrollFactorY = child.scrollFactorY;
 
-        //  Set new values
+        //  Set parent values
         child.setScrollFactor(childScrollFactorX * scrollFactorX, childScrollFactorY * scrollFactorY);
         child.setAlpha(childAlpha * alpha);
+
+        if (containerHasBlendMode)
+        {
+            child.setBlendMode(container._blendMode);
+        }
 
         //  Render
         child.renderCanvas(renderer, child, interpolationPercentage, camera, transformMatrix);
 
-        //  Restore old values
+        //  Restore original values
         child.setAlpha(childAlpha);
         child.setScrollFactor(childScrollFactorX, childScrollFactorY);
+        child.setBlendMode(childBlendMode);
     }
 };